backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
google-genai>=0.3
pydantic>=2
python-dotenv>=1.0
